@@ -7,7 +7,7 @@ using static System.StringComparison;
 
 namespace Todo.Helpers
 {
-    public static class IQueryableExtensions
+    public static class QueryableExtensions
     {
         public static IQueryable<T> ApplySort<T>(this IQueryable<T> source, string orderBy, IDictionary<string, PropertyMappingValue> propertyMapping)
         {
