@@ -2,9 +2,9 @@
 {
     public class LinkDto
     {
-        public string Href { get; set; }
-        public string Rel { get; set; }
-        public string Method { get; set; }
+        public string Href { get; }
+        public string Rel { get; }
+        public string Method { get; }
 
         public LinkDto() { }
 
