@@ -4,11 +4,11 @@ namespace Todo.Models
 {
     public class TodoItemFullDto
     {
-        public Guid Id { get; set; }
-        public string Name { get; set; }
-        public string Project { get; set; }
-        public string Context { get; set; }
-        public DateTime Date { get; set; }
-        public bool IsComplete { get; set; }
+        public Guid Id { get; init; }
+        public string Name { get; init; }
+        public string Project { get; init; }
+        public string Context { get; init; }
+        public DateTime Date { get; init; }
+        public bool IsComplete { get; init; }
     }
 }
