@@ -5,6 +5,6 @@ namespace Todo.Models
     public class TodoItemUpdateDto : TodoItemManipulationDto
     {
         [Required]
-        public override string Project { get => base.Project; set => base.Project = value; }
+        public override string Project { get => base.Project; init => base.Project = value; }
     }
 }
