@@ -1,7 +1,4 @@
-﻿using System.Diagnostics.CodeAnalysis;
-
-namespace Todo.Services
+﻿namespace Todo.Services
 {
-    [SuppressMessage("Design", "CA1040:Avoid empty interfaces", Justification = "<Pending>")]
     public interface IPropertyMapping { }
 }
